refactor template methods
